Code organization using ARP & utils
